backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.7.0
python-multipart==0.0.20
argon2-cffi==23.1.0
itsdangerous==2.2.0
aiofiles==24.1.0
pyyaml==6.0.2
httpx==0.28.0
psutil==6.1.0
pytest==8.3.0
pytest-asyncio==0.25.0
